backend4/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
